Remove redundant support files
- Prune unused helpers
- Inline single use helpers
- Remove any cases that rely on inferring visual behaviour
  already covered by backstop tests
