+++ testing/.env.example
@@ -1,11 +1,5 @@
-# Browser width (If visible)
-BROWSER_WIDTH=1280
-
-# Browser height (If visible)
-BROWSER_HEIGHT=800
-
 # Use headless browser. Defaults to true
-HEADLESS=true
+# HEADLESS=false
 
 # The base url of the Design System website you are hitting
-TESTING_BASE_URL=https://citizensadvice.github.io/design-system
+TESTING_BASE_URL=http://localhost:6006/
