+++ testing/cucumber.yml
@@ -1,3 +1,3 @@
-passing: --tags 'not @failing and not @future_release'
+passing: --tags 'not @failing'
 
 reports: --format progress -p passing
