--- testing/features/support/all.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-require "capybara"
-require "capybara/dsl"
-require "capybara/cucumber"
-require "forwardable"
-require "logger"
-require "rspec"
-require "selenium-webdriver"
-require "singleton"
-require "site_prism"
-require "webdrivers"
