+++ testing/features/support/components/asset_hyperlink/asset_hyperlink.rb
@@ -2,16 +2,10 @@
 
 module AssetHyperlink
   class Example < ::Base
-    include Helpers::Javascript
-
     set_url "/iframe.html?id=components-asset-hyperlink--example&viewMode=story"
 
     element :download_link, "a"
     element :download_size, ".cads-asset-type"
-
-    def download_icon_content
-      before_content(".cads-icon_file").delete('\\"')
-    end
 
     def validate_initial_state!
       has_download_link?(wait: 5)
