--- testing/features/support/helpers/all.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "env_variables"
-
-require_relative "browser_window"
-require_relative "javascript"
-require_relative "regex"
