+++ testing/features/support/hooks.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
 Before do |test_case|
-  test_case.source_tag_names.include?("@small_screen") ? resize_window(320, 568) : resize_window
+  if test_case.source_tag_names.include?("@small_screen")
+    Capybara.current_window.resize_to(320, 568)
+  else
+    Capybara.current_window.resize_to(1280, 800)
+  end
 end
